pyproject.toml
[tool.poetry]
name = "frontend-test-crew"
version = "0.1.0"
description = "Multi-agent frontend testing system using CrewAI and Playwright"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
crewai = ">=0.28.0"
crewai-tools = ">=0.1.0"
playwright = ">=1.40.0"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
